Move the test into a new file
